models/advanced_features.py
```python
from odoo import models, fields, api

class AdvancedFeatures(models.Model):
    _name = 'advanced.features'
    _description = 'Advanced Features'

    name = fields.Char(string='Feature Name', required=True)
    description = fields.Text(string='Description')
    is_active = fields.Boolean(string='Is Active?', default=True)
    feature_type = fields.Selection([
        ('3d_visualization', '3D Visualization'),
        ('performance_estimation', 'Performance Estimation'),
        # Add more feature types as needed
    ], string='Feature Type', required=True)

    @api.onchange('feature_type')
    def _onchange_feature_type(self):
        # Implement business logic based on feature type
        pass

    def toggle_feature(self):
        self.is_active = not self.is_active

    def integrate_third_party_services(self):
        # Implement integration with third-party libraries or services as needed
        pass
```